requirements.txt
requests
python-whois
python-tld
